Move defaulting to a separate go file
